requirements.txt
# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Chemistry and molecular libraries
rdkit-pypi>=2022.9.1
datamol>=0.9.0
molfeat>=0.5.0
mols2grid>=1.0.0

# Machine learning and deep learning
tensorflow>=2.8.0
torch>=1.12.0
transformers>=4.20.0

# Molecular generation and similarity
PyTDC>=0.4.0
safe-mol>=0.1.0

# Vector database
pinecone-client>=2.2.0

# Data visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Progress bars and utilities
tqdm>=4.64.0

# Configuration and logging
pyyaml>=6.0
python-dotenv>=0.19.0

# COATI models (if available)
# coati-models>=0.1.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950 